fix: address PR #164 feedback — extract browser-safe Strava URL helper, centralize colors
- Extract buildStravaUrl/extractStravaActivityId into src/lib/strava/url.ts
  (browser-safe, no prisma/pg dependency — fixes Vercel build failure from edc3126)
- Re-export from client.ts so existing server consumers are unaffected
- Add --color-strava and --color-strava-hover to Tailwind @theme inline block
- Replace all hardcoded #FC4C02/#E34402 hex values across 4 components
- Remove redundant stopPropagation on sibling <a> tag in EditAttendanceDialog
- Fix pending-state UX: disable hover and external link during mutations
